# api/swagger.yaml
openapi: 3.0.3
info:
  title: wlachs' blog
  description: |-
    This is a lightweight blog engine implemented in Go.
  contact:
    email: hello@laszloborbely.com
  license:
    name: MIT
    url: https://github.com/wlachs/blog/blob/main/LICENSE
  version: 0.1.0
servers:
  - url: https://laszloborbely.com/api/v0
...
paths:
...
  /users/{UserID}:
    parameters:
      - $ref: '#/components/parameters/UserID'
    get:
      tags:
        - User
      summary: Get user by ID
      description: Find and retrieve user by the given ID
      operationId: getUserByID
      responses:
        200:
          description: Successfully retrieved user by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User with the provided ID not found
    post:
      tags:
        - User
      summary: Add new user
      description: Add a new user to the blog with posting rights
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Password object
              required:
                - password
              properties:
                password:
                  type: string
                  description: User password
                  format: password
                  example: '*****'
      responses:
        201:
          description: Successfully added new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Missing credentials
        409:
          description: User with the provided ID already exists
      security:
        - X-Auth-Token: [ ]
    put:
      tags:
        - User
      summary: Change existing user
      description: Change password of an existing user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Password change object
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  type: string
                  description: Old user password
                  format: password
                  example: Test1234
                newPassword:
                  type: string
                  description: New user password
                  format: password
                  example: Test1234
      responses:
        200:
          description: Successfully updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Incorrect user name or password
      security:
        - X-Auth-Token: [ ]
...
components:
  parameters:
...
    UserID:
      name: UserID
      description: Unique user identifier
      in: path
      required: true
      schema:
        type: string
  schemas:
    PostMetadata:
      type: object
      description: Contains post metadata without the body
      required:
        - id
        - title
        - author
        - creationTime
      properties:
        id:
          type: string
          description: Unique post identifier
          example: interesting-post-title-in-url
        title:
          type: string
          description: Post title
          example: Interesting Post
        author:
          type: string
          description: Name of the author
          example: Laszlo
        summary:
          type: string
          description: Short summary of the post. Typically not longer than a few sentences
          example: Interesting Post Summary
        creationTime:
          type: string
          format: date-time
          description: Date when the post was created
          example: "2023-11-21T22:55:30.335Z"
...
    User:
      type: object
      description: Object representing a blog user
      required:
        - userID
      properties:
        userID:
          type: string
          description: Unique user identifier
          example: Laszlo
        posts:
          type: array
          description: Posts authored by the user
          items:
            $ref: '#/components/schemas/PostMetadata'
...
  securitySchemes:
    X-Auth-Token:
      type: apiKey
      name: X-Auth-Token
      in: header